+++ ConvolutionalNeuralNetwork/DigitImageLoadingService.h
@@ -8,5 +8,5 @@ typedef unsigned char uchar;
 static class DigitImageLoadingService {
 public:
 	static MatrixBlock* read_mnist_images(std::string full_path, int& number_of_images);
-	static float* read_mnist_labels(std::string full_path, int& number_of_labels);
+	static int* read_mnist_labels(std::string full_path, int& number_of_labels);
 };
